+++ commands/run/__tests__/run-command.test.js
@@ -331,6 +331,13 @@ describe("RunCommand", () => {
       expect(collectedOutput).toContain("Successfully ran target");
     });
 
+    it("runs a script with a colon in the script name", async () => {
+      collectedOutput = "";
+      await lernaRun(testDir)("another-script:but-with-colons");
+      expect(collectedOutput).toContain("package-1-script-with-colons");
+      expect(collectedOutput).toContain("Successfully ran target");
+    });
+
     it("runs a script only in scoped packages", async () => {
       collectedOutput = "";
       await lernaRun(testDir)("my-script", "--scope", "package-1");