+++ e2e/tests/lerna-run/lerna-run.spec.ts
@@ -2,8 +2,6 @@ import { existsSync } from "fs-extra";
 import { Fixture } from "../../utils/fixture";
 import { normalizeCommandOutput, normalizeEnvironment } from "../../utils/snapshot-serializer-utils";
 
-jest.setTimeout(60000);
-
 expect.addSnapshotSerializer({
   serialize(str: string) {
     return normalizeCommandOutput(normalizeEnvironment(str));
@@ -13,7 +11,7 @@ expect.addSnapshotSerializer({
   },
 });
 
-describe("lerna run", () => {
+describe("lerna-run", () => {
   let fixture: Fixture;
 
   beforeAll(async () => {
@@ -312,6 +310,22 @@ describe("useNx", () => {
         "print-name": "echo test-package-3",
       },
     });
+    await fixture.lerna("create package-4 -y");
+    await fixture.addScriptsToPackage({
+      packagePath: "packages/package-4",
+      scripts: {
+        "print:name": "echo test-package-4",
+        "print-name-run-one-only": "echo test-package-4-run-one-only",
+        "print:name:run-one-only": "echo test-package-4-run-one-only-with-colon",
+      },
+    });
+    await fixture.lerna("create package-5 -y");
+    await fixture.addScriptsToPackage({
+      packagePath: "packages/package-5",
+      scripts: {
+        "print:name": "echo test-package-5",
+      },
+    });
   });
   afterAll(() => fixture.destroy());
 
@@ -355,6 +369,92 @@ test-package-X
  
 
  >  Lerna (powered by Nx)   Successfully ran target print-name for 3 projects
+
+
+lerna notice cli v999.9.9-e2e.0
+
+`);
+  });
+
+  describe("run one", () => {
+    it("should run script on single child package using nx", async () => {
+      const output = await fixture.lerna(`run print-name-run-one-only`);
+
+      expect(output.combinedOutput).toMatchInlineSnapshot(`
+  
+  > package-X:print-name-run-one-only
+  
+  
+  > package-X@0.0.0 print-name-run-one-only
+  > echo test-package-X-run-one-only
+  
+  test-package-X-run-one-only
+  
+   
+  
+   >  Lerna (powered by Nx)   Successfully ran target print-name-run-one-only for project package-X
+  
+  
+  lerna notice cli v999.9.9-e2e.0
+  
+  `);
+    });
+
+    it("should run script with colon on single child package using nx", async () => {
+      const output = await fixture.lerna(`run print:name:run-one-only`);
+
+      expect(output.combinedOutput).toMatchInlineSnapshot(`
+
+  > package-X:"print:name:run-one-only"
+  
+  
+  > package-X@0.0.0 print:name:run-one-only
+  > echo test-package-X-run-one-only-with-colon
+  
+  test-package-X-run-one-only-with-colon
+  
+   
+  
+   >  Lerna (powered by Nx)   Successfully ran target print:name:run-one-only for project package-X
+  
+  
+  lerna notice cli v999.9.9-e2e.0
+  
+  `);
+    });
+  });
+
+  it("should run script with colon on all child package using nx", async () => {
+    const output = await fixture.lerna(`run print:name`);
+
+    expect(output.combinedOutput).toMatchInlineSnapshot(`
+
+ >  Lerna (powered by Nx)   Running target print:name for 2 project(s):
+
+    - package-X
+    - package-X
+
+ 
+
+> package-X:"print:name"
+
+
+> package-X@0.0.0 print:name
+> echo test-package-X
+
+test-package-X
+
+> package-X:"print:name"
+
+
+> package-X@0.0.0 print:name
+> echo test-package-X
+
+test-package-X
+
+ 
+
+ >  Lerna (powered by Nx)   Successfully ran target print:name for 2 projects
 
 
 lerna notice cli v999.9.9-e2e.0